requirements.txt
# requirements.txt
requests>=2.28.0
googlemaps>=4.7.0
pandas>=1.5.0
python-dotenv>=0.19.0
pymongo>=4.3.0
pytest>=7.0.0
pytest-cov>=4.0.0